--- examples/rsx_autocomplete.rs
@@ -1,27 +1,0 @@
-//! This example shows that autocomplete works in RSX
-
-use dioxus::prelude::*;
-
-fn main() {
-    dioxus::desktop::launch(app);
-}
-
-fn app(cx: Scope) -> Element {
-    cx.render(rsx! {
-        div {
-            onclick: move |_| {
-            }
-            // class: "asd",
-            // style {
-            //     media: "Ad",
-            // }
-            // div {
-
-            // }
-            // {
-            //     let t = String::new();
-            //     t.
-            // }
-        }
-    })
-}
