+++ packages/rsx/src/pretty.rs
@@ -1,0 +1,1 @@
+//! pretty printer for rsx!
